requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
affine=2.4.0=pyhd8ed1ab_0
anyio=3.6.2=pypi_0
aom=3.5.0=h7ea286d_0
appdirs=1.4.4=pyh9f0ad1d_0
appnope=0.1.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
attrs=22.2.0=pyh71513ae_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.2=pypi_0
black=23.3.0=py39h2804cbe_1
bleach=6.0.0=pypi_0
blosc=1.21.3=h1d6ff8b_0
boost-cpp=1.78.0=h1cb353e_1
branca=0.6.0=pyhd8ed1ab_0
brotli=1.0.9=h1a8c8d9_8
brotli-bin=1.0.9=h1a8c8d9_8
brotlipy=0.7.0=py39h02fc5c5_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2023.5.7=hf0a4a13_0
cachetools=5.3.0=pyhd8ed1ab_0
cairo=1.16.0=h73a0509_1014
certifi=2022.12.7=pypi_0
cffi=1.15.1=py39h7e6b969_3
cfitsio=4.2.0=h2f961c4_0
cftime=1.6.2=pypi_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=unix_pyhd8ed1ab_2
click-plugins=1.1.1=py_0
cligj=0.7.2=pyhd8ed1ab_1
comm=0.1.2=pypi_0
contourpy=1.0.7=py39haaf3ac1_0
cryptography=39.0.1=py39he2a39a8_0
curl=7.88.1=h9049daf_1
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.6.6=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pypi_0
executing=1.2.0=pypi_0
expat=2.5.0=hb7217d7_0
fastjsonschema=2.16.2=pypi_0
ffmpeg=5.1.2=gpl_hf318d42_106
fiona=1.9.1=py39h9e5269e_0
folium=0.14.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h82840c6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=hd633e50_1
freexl=1.0.6=h1a8c8d9_1
gdal=3.6.2=py39h766d3fc_9
geocube=0.4.0=pyhd8ed1ab_0
geopandas=0.12.2=pyhd8ed1ab_0
geopandas-base=0.12.2=pyha770c72_0
geos=3.11.1=hb7217d7_0
geotiff=1.7.1=hdcdc974_6
gettext=0.21.1=h0186832_0
giflib=5.2.1=h1a8c8d9_3
gmp=6.2.1=h9f76cd9_0
gnutls=3.7.8=h9f1a10d_0
h5netcdf=1.1.0=pypi_0
h5py=3.8.0=pypi_0
hdf4=4.2.15=h1a38d6a_5
hdf5=1.12.2=nompi_ha7af310_101
icu=70.1=h6b3803e_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.12.0=pyhd8ed1ab_0
importlib_resources=5.12.0=pyhd8ed1ab_0
ipykernel=6.21.1=pypi_0
ipython=8.10.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.4=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h1a8c8d9_3
json-c=0.16=hc449e50_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter-client=8.0.2=pypi_0
jupyter-core=5.2.0=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.2.1=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.5=pypi_0
kealib=1.5.0=hfd766a6_0
kiwisolver=1.4.4=py39haaf3ac1_1
krb5=1.20.1=h69eda48_0
lame=3.100=h1a8c8d9_1003
lcms2=2.14=h481adae_1
lerc=4.0.0=h9a09cb3_0
libaec=1.0.6=hb7217d7_1
libblas=3.9.0=16_osxarm64_openblas
libbrotlicommon=1.0.9=h1a8c8d9_8
libbrotlidec=1.0.9=h1a8c8d9_8
libbrotlienc=1.0.9=h1a8c8d9_8
libcblas=3.9.0=16_osxarm64_openblas
libcurl=7.88.1=h9049daf_1
libcxx=14.0.6=h2692d47_0
libdeflate=1.17=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgdal=3.6.2=h8d4b95d_9
libgfortran=5.0.0=11_3_0_hd922786_27
libgfortran5=11.3.0=hdaf2cc0_27
libglib=2.74.1=h4646484_1
libiconv=1.17=he4db4b2_0
libidn2=2.3.4=h1a8c8d9_0
libkml=1.3.0=h41464e4_1015
liblapack=3.9.0=16_osxarm64_openblas
libnetcdf=4.9.1=nompi_h232cb48_101
libnghttp2=1.52.0=hae82a92_0
libopenblas=0.3.21=openmp_hc731615_3
libopus=1.3.1=h27ca646_1
libpng=1.6.39=h76d750c_0
libpq=15.2=h1a28acd_0
librttopo=1.1.0=h844f84d_12
libspatialindex=1.9.3=hbdafb3b_4
libspatialite=5.0.1=h14115fc_23
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=h7a5bd25_3
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.5.0=h5dffbdd_2
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.10.3=h87b0503_0
libzip=1.9.2=h76ab92c_1
libzlib=1.2.13=h03a7124_4
llvm-openmp=15.0.7=h7cfbb63_0
lz4-c=1.9.4=hb7217d7_0
mapclassify=2.5.0=pyhd8ed1ab_1
markupsafe=2.1.2=py39h02fc5c5_0
matplotlib=3.6.3=pypi_0
matplotlib-base=3.7.1=py39h35e9e80_0
matplotlib-inline=0.1.6=pypi_0
mistune=2.0.5=pypi_0
munch=2.5.0=py_0
munkres=1.1.4=pyh9f0ad1d_0
mypy_extensions=1.0.0=pyha770c72_0
nbclassic=0.5.1=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.9=pypi_0
nbformat=5.7.3=pypi_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pypi_0
netcdf4=1.6.3=pypi_0
nettle=3.8.1=h63371fa_1
networkx=3.0=pyhd8ed1ab_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
nspr=4.35=hb7217d7_0
nss=3.89=h789eff7_0
numpy=1.24.2=py39hff61c6a_0
odc-geo=0.3.3=pyhd8ed1ab_0
opencv-python=3.4.18.65=pypi_0
openeo=0.14.1=pypi_0
openh264=2.3.1=hb7217d7_2
openjpeg=2.5.0=hbc2ba62_2
openssl=3.1.0=h53f4e23_3
p11-kit=0.24.1=h29577a5_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.0=pyhd8ed1ab_0
pcre2=10.40=hb34f9b4_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py39h8bd98a6_1
pip=23.0=pyhd8ed1ab_0
pixman=0.40.0=h27ca646_0
platformdirs=3.0.0=pyhd8ed1ab_0
pooch=1.6.0=pyhd8ed1ab_0
poppler=23.03.0=h9564b9f_0
poppler-data=0.4.12=hd8ed1ab_0
postgresql=15.2=h45c140d_0
proj=9.1.1=h13f728c_2
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.36=pypi_0
psutil=5.9.4=pypi_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.14.0=pypi_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyproj=3.4.1=py39hbb572e0_1
pyrsistent=0.19.3=pypi_0
pyshp=2.3.1=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.16=hea58f1e_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-json-logger=2.0.5=pypi_0
python_abi=3.9=3_cp39
pytorch=1.13.1=py3.9_0
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.0=pypi_0
rasterio=1.3.5.post1=pypi_0
readline=8.1.2=h46ed386_0
requests=2.28.2=pyhd8ed1ab_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rioxarray=0.14.0=pyhd8ed1ab_1
rtree=1.0.1=py39hb28b0e7_1
scikit-learn=1.2.2=py39h6180588_1
scipy=1.10.0=py39h18313fe_2
send2trash=1.8.0=pypi_0
setuptools=67.1.0=pyhd8ed1ab_0
shapely=2.0.1=py39h472ea82_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=h17c5cce_0
sniffio=1.3.0=pypi_0
snuggs=1.4.7=py_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.40.0=h2229b38_0
stack-data=0.6.2=pypi_0
svt-av1=1.4.1=h7ea286d_0
terminado=0.17.1=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2023.3.21=pypi_0
tiledb=2.13.2=h9bd36d0_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=he1e0b03_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=0.13.1=py39_cpu
torchvision=0.14.1=py39_cpu
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzcode=2023b=h1a8c8d9_0
tzdata=2022g=h191b570_0
unicodedata2=15.0.0=py39h02fc5c5_0
uri-template=1.2.0=pypi_0
urllib3=1.26.14=pyhd8ed1ab_0
wcwidth=0.2.6=pypi_0
webcolors=1.12=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
widgetsnbextension=4.0.5=pypi_0
wrapt=1.14.1=pypi_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xarray=2023.2.0=pypi_0
xerces-c=3.2.4=h627aa08_1
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xyzservices=2023.2.0=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
zipp=3.13.0=pypi_0
zlib=1.2.13=h03a7124_4
zstd=1.5.2=hf913c23_6
